Return process.env on Windows
On Windows it is not expected to export vars in .bash_profile, .bashrc, etc. So we should always return the `process.env` variables.
As I have to add unit tests for the functionality, but I'm unable to modify process.platform, I've added process-wrapper that will be used for this purpose.