Get variables from Node.js directly
Parsing stdout of a process could fail due to many reasons. But executing Node.js in a child process, that already has all required variables, will allow us to get the `process.env` object. It already has parsed the variables and we just have to read them.
So introduce this as a first option, in case there's Node.js on user's machine.