Support custom user configs for shell and terminal
In many cases users may have different terminals/shells, so the default settings may not work on their side.
So make it possible to specify the default shell and the type of terminal - is it login and is it interactive.
Pass object to `getEnvironmentVariables` that has the following type:
```TypeScript
{
    shell: string;
    terminalConfiguration: {
        isInteractive: boolean;
        isLogin: boolean;
    }
}
```

The default values are:
- shell will be set to `process.env.SHELL` or `/bin/bash`.
- `isInteractive` will be `true`
- `isLogin` will be false for Linux and true for all other OSes.

NOTE: This is breaking change for Linux users - until now we've spawned login shell.